+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 -   Adds support for `.js` data files ([#42](https://github.com/marceljs/marcel/issues/42)).
+-   Adds `build` (default) and `serve` commands to the CLI ([#22](https://github.com/marceljs/marcel/issues/22)).
 
 ### 0.1.3